+++ create.sh
@@ -7,6 +7,6 @@ docker rmi -f petrjahoda/system_service:latest
 docker build -t petrjahoda/system_service:latest .
 docker push petrjahoda/system_service:latest
 
-docker rmi -f petrjahoda/system_service:2020.3.1
-docker build -t petrjahoda/system_service:2020.3.1 .
-docker push petrjahoda/system_service:2020.3.1
+docker rmi -f petrjahoda/system_service:2020.3.2
+docker build -t petrjahoda/system_service:2020.3.2 .
+docker push petrjahoda/system_service:2020.3.2